# frontend/openapi.json
{
	"components": {
		"schemas": {
			"APIErrorBody": {
				"properties": {
					"code": {
						"title": "Code",
						"type": "string"
					},
					"message": {
						"title": "Message",
						"type": "string"
					},
					"request_id": {
						"title": "Request Id",
						"type": "string"
					},
					"retryable": {
						"default": false,
						"title": "Retryable",
						"type": "boolean"
					}
				},
				"required": ["code", "message", "request_id"],
				"title": "APIErrorBody",
				"type": "object"
			},
...
			"PaginatedResponse_VideoOut_": {
				"properties": {
					"has_more": {
						"title": "Has More",
						"type": "boolean"
					},
					"items": {
						"items": {
							"$ref": "#/components/schemas/VideoOut"
						},
						"title": "Items",
						"type": "array"
					},
					"limit": {
						"title": "Limit",
						"type": "integer"
					},
					"offset": {
						"title": "Offset",
						"type": "integer"
					},
					"total": {
						"title": "Total",
						"type": "integer"
					}
				},
				"required": ["total", "items", "limit", "offset", "has_more"],
				"title": "PaginatedResponse[VideoOut]",
				"type": "object"
			},
			"PlaylistAddVideo": {
				"description": "Schema for adding a video to a playlist.",
				"properties": {
					"position": {
						"anyOf": [
							{
								"minimum": 0.0,
								"type": "integer"
							},
							{
								"type": "null"
							}
						],
						"description": "Position in playlist (appended if omitted)",
						"title": "Position"
					},
					"video_id": {
						"description": "Video ID to add",
						"title": "Video Id",
						"type": "string"
					}
				},
				"required": ["video_id"],
				"title": "PlaylistAddVideo",
				"type": "object"
			},
...
			"PlaylistDetailOut": {
				"description": "Schema for detailed playlist output with ordered video IDs.",
				"properties": {
					"created_at": {
						"format": "date-time",
						"title": "Created At",
						"type": "string"
					},
					"current_position": {
						"anyOf": [
							{
								"type": "integer"
							},
							{
								"type": "null"
							}
						],
						"title": "Current Position"
					},
					"description": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Description"
					},
					"id": {
						"title": "Id",
						"type": "string"
					},
					"is_system": {
						"title": "Is System",
						"type": "boolean"
					},
					"name": {
						"title": "Name",
						"type": "string"
					},
					"source_channel_id": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Source Channel Id"
					},
					"source_is_active": {
						"title": "Source Is Active",
						"type": "boolean"
					},
					"source_last_synced_at": {
						"anyOf": [
							{
								"format": "date-time",
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Source Last Synced At"
					},
					"source_type": {
						"title": "Source Type",
						"type": "string"
					},
					"source_youtube_playlist_id": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Source Youtube Playlist Id"
					},
					"thumbnail_url": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Thumbnail Url"
					},
					"total_videos": {
						"title": "Total Videos",
						"type": "integer"
					},
					"video_ids": {
						"items": {
							"type": "string"
						},
						"title": "Video Ids",
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"description",
					"is_system",
					"source_type",
					"source_is_active",
					"current_position",
					"total_videos",
					"created_at",
					"video_ids"
				],
				"title": "PlaylistDetailOut",
				"type": "object"
			},
...
			"PlaylistSetVideos": {
				"description": "Schema for setting the full ordered video list of a playlist.",
				"properties": {
					"video_ids": {
						"description": "Ordered list of video IDs",
						"items": {
							"type": "string"
						},
						"title": "Video Ids",
						"type": "array"
					}
				},
				"required": ["video_ids"],
				"title": "PlaylistSetVideos",
				"type": "object"
			},
...
			"VideoOut": {
				"description": "Schema for returning a video from the API.",
				"properties": {
					"channel_id": {
						"title": "Channel Id",
						"type": "string"
					},
					"created_at": {
						"format": "date-time",
						"title": "Created At",
						"type": "string"
					},
					"description": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Description"
					},
					"duration_seconds": {
						"anyOf": [
							{
								"type": "integer"
							},
							{
								"type": "null"
							}
						],
						"title": "Duration Seconds"
					},
					"id": {
						"title": "Id",
						"type": "string"
					},
					"is_favorited": {
						"title": "Is Favorited",
						"type": "boolean"
					},
					"is_short": {
						"title": "Is Short",
						"type": "boolean"
					},
					"is_watched": {
						"title": "Is Watched",
						"type": "boolean"
					},
					"published_at": {
						"format": "date-time",
						"title": "Published At",
						"type": "string"
					},
					"thumbnail_url": {
						"anyOf": [
							{
								"format": "uri",
								"maxLength": 2083,
								"minLength": 1,
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Thumbnail Url"
					},
					"title": {
						"title": "Title",
						"type": "string"
					},
					"yt_tags": {
						"items": {
							"type": "string"
						},
						"title": "Yt Tags",
						"type": "array"
					}
				},
				"required": [
					"id",
					"channel_id",
					"title",
					"description",
					"thumbnail_url",
					"published_at",
					"duration_seconds",
					"yt_tags",
					"is_short",
					"is_favorited",
					"is_watched",
					"created_at"
				],
				"title": "VideoOut",
				"type": "object"
			},
...
		},
		"securitySchemes": {
			"OAuth2PasswordBearer": {
				"flows": {
					"password": {
						"scopes": {},
						"tokenUrl": "auth/jwt/login"
					}
				},
				"type": "oauth2"
			}
		}
	},
	"info": {
		"license": {
			"identifier": "GPL-3.0-only",
			"name": "GPL-3.0-only"
		},
		"summary": "Backend API for the ChooseYourTube distraction-free YouTube reader.",
		"title": "ChooseYourTube API",
		"version": "0.1.0"
	},
	"openapi": "3.1.0",
	"paths": {
...
		"/playlists/{playlist_id}/videos": {
			"delete": {
				"description": "Clear all videos from a playlist.",
				"operationId": "clear_playlist_videos_playlists__playlist_id__videos_delete",
				"parameters": [
					{
						"in": "path",
						"name": "playlist_id",
						"required": true,
						"schema": {
							"title": "Playlist Id",
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PlaylistDetailOut"
								}
							}
						},
						"description": "Successful Response"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"403": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"409": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"422": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					}
				},
				"security": [
					{
						"OAuth2PasswordBearer": []
					}
				],
				"summary": "Clear Playlist Videos",
				"tags": ["Playlists"]
			},
			"post": {
				"description": "Add a video to a playlist. If already present, moves it to the requested position.",
				"operationId": "add_video_to_playlist_playlists__playlist_id__videos_post",
				"parameters": [
					{
						"in": "path",
						"name": "playlist_id",
						"required": true,
						"schema": {
							"title": "Playlist Id",
							"type": "string"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PlaylistAddVideo"
							}
						}
					},
					"required": true
				},
				"responses": {
					"201": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PlaylistDetailOut"
								}
							}
						},
						"description": "Successful Response"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"403": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"409": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"422": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					}
				},
				"security": [
					{
						"OAuth2PasswordBearer": []
					}
				],
				"summary": "Add Video To Playlist",
				"tags": ["Playlists"]
			},
			"put": {
				"description": "Set the full ordered video list for a playlist.",
				"operationId": "set_playlist_videos_playlists__playlist_id__videos_put",
				"parameters": [
					{
						"in": "path",
						"name": "playlist_id",
						"required": true,
						"schema": {
							"title": "Playlist Id",
							"type": "string"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PlaylistSetVideos"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PlaylistDetailOut"
								}
							}
						},
						"description": "Successful Response"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"403": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"409": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"422": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					}
				},
				"security": [
					{
						"OAuth2PasswordBearer": []
					}
				],
				"summary": "Set Playlist Videos",
				"tags": ["Playlists"]
			}
		},
...
		"/playlists/{playlist_id}/videos/{video_id}": {
			"delete": {
				"description": "Remove a video from a playlist.",
				"operationId": "remove_video_from_playlist_playlists__playlist_id__videos__video_id__delete",
				"parameters": [
					{
						"in": "path",
						"name": "playlist_id",
						"required": true,
						"schema": {
							"title": "Playlist Id",
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "video_id",
						"required": true,
						"schema": {
							"title": "Video Id",
							"type": "string"
						}
					}
				],
				"responses": {
					"204": {
						"description": "Successful Response"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"403": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"409": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"422": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					}
				},
				"security": [
					{
						"OAuth2PasswordBearer": []
					}
				],
				"summary": "Remove Video From Playlist",
				"tags": ["Playlists"]
			}
		},
...
		"/tags/{tag_id}/videos": {
			"get": {
				"description": "Get all videos associated with a tag.",
				"operationId": "get_videos_for_tag_tags__tag_id__videos_get",
				"parameters": [
					{
						"in": "path",
						"name": "tag_id",
						"required": true,
						"schema": {
							"title": "Tag Id",
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 50,
							"maximum": 200,
							"minimum": 1,
							"title": "Limit",
							"type": "integer"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"default": 0,
							"minimum": 0,
							"title": "Offset",
							"type": "integer"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/VideoOut"
									},
									"title": "Response Get Videos For Tag Tags  Tag Id  Videos Get",
									"type": "array"
								}
							}
						},
						"description": "Successful Response"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"403": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"409": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"422": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					}
				},
				"security": [
					{
						"OAuth2PasswordBearer": []
					}
				],
				"summary": "Get Videos For Tag",
				"tags": ["Tags"]
			}
		},
...
		"/videos/": {
			"get": {
				"description": "List videos with optional filtering and search.\n\nFilters:\n- is_favorited: Show only favorited videos\n- is_watched: Show only watched/unwatched videos\n- is_short: Show only YouTube Shorts or exclude them\n- channel_id: Filter by single channel ID or multiple IDs (comma-separated)\n- tag_id: Show only videos with a specific tag\n- published_after: Show videos published after a date (e.g., \"2026-01-01T00:00:00Z\")\n- published_before: Show videos published before a date (e.g., \"2026-12-31T23:59:59Z\")\n- q: Full-text search across video title, description, and tag names\n- order_by: Sort field (default: published_at; use \"relevance\" with search queries)\n- order_direction: Sort direction (asc or desc)",
				"operationId": "list_videos_videos__get",
				"parameters": [
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 50,
							"maximum": 200,
							"minimum": 1,
							"title": "Limit",
							"type": "integer"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"default": 0,
							"minimum": 0,
							"title": "Offset",
							"type": "integer"
						}
					},
					{
						"description": "Filter by favorited status",
						"in": "query",
						"name": "is_favorited",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"type": "boolean"
								},
								{
									"type": "null"
								}
							],
							"description": "Filter by favorited status",
							"title": "Is Favorited"
						}
					},
					{
						"description": "Filter by watched status",
						"in": "query",
						"name": "is_watched",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"type": "boolean"
								},
								{
									"type": "null"
								}
							],
							"description": "Filter by watched status",
							"title": "Is Watched"
						}
					},
					{
						"description": "Filter by YouTube Shorts",
						"in": "query",
						"name": "is_short",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"type": "boolean"
								},
								{
									"type": "null"
								}
							],
							"description": "Filter by YouTube Shorts",
							"title": "Is Short"
						}
					},
					{
						"description": "Filter by channel ID (single ID or comma-separated list)",
						"in": "query",
						"name": "channel_id",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "null"
								}
							],
							"description": "Filter by channel ID (single ID or comma-separated list)",
							"examples": ["ch001", "ch001,ch002,ch003"],
							"title": "Channel Id"
						}
					},
					{
						"description": "Filter by tag ID",
						"in": "query",
						"name": "tag_id",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "null"
								}
							],
							"description": "Filter by tag ID",
							"title": "Tag Id"
						}
					},
					{
						"description": "Filter videos published after this date (ISO 8601 format)",
						"in": "query",
						"name": "published_after",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "null"
								}
							],
							"description": "Filter videos published after this date (ISO 8601 format)",
							"title": "Published After"
						}
					},
					{
						"description": "Filter videos published before this date (ISO 8601 format)",
						"in": "query",
						"name": "published_before",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "null"
								}
							],
							"description": "Filter videos published before this date (ISO 8601 format)",
							"title": "Published Before"
						}
					},
					{
						"description": "Search videos by title, description, or tag name",
						"in": "query",
						"name": "q",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "null"
								}
							],
							"description": "Search videos by title, description, or tag name",
							"title": "Q"
						}
					},
					{
						"description": "Sort field: published_at, title, created_at, duration_seconds, relevance (with q)",
						"in": "query",
						"name": "order_by",
						"required": false,
						"schema": {
							"default": "published_at",
							"description": "Sort field: published_at, title, created_at, duration_seconds, relevance (with q)",
							"title": "Order By",
							"type": "string"
						}
					},
					{
						"description": "Sort direction",
						"in": "query",
						"name": "order_direction",
						"required": false,
						"schema": {
							"default": "desc",
							"description": "Sort direction",
							"enum": ["asc", "desc"],
							"title": "Order Direction",
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaginatedResponse_VideoOut_"
								}
							}
						},
						"description": "Successful Response"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"403": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"409": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"422": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					}
				},
				"security": [
					{
						"OAuth2PasswordBearer": []
					}
				],
				"summary": "List Videos",
				"tags": ["Videos"]
			}
		},
		"/videos/by-channel/{channel_id}": {
			"get": {
				"operationId": "list_videos_by_channel_videos_by_channel__channel_id__get",
				"parameters": [
					{
						"in": "path",
						"name": "channel_id",
						"required": true,
						"schema": {
							"title": "Channel Id",
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 50,
							"maximum": 200,
							"minimum": 1,
							"title": "Limit",
							"type": "integer"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"default": 0,
							"minimum": 0,
							"title": "Offset",
							"type": "integer"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/VideoOut"
									},
									"title": "Response List Videos By Channel Videos By Channel  Channel Id  Get",
									"type": "array"
								}
							}
						},
						"description": "Successful Response"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"403": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"409": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"422": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIErrorBody"
								}
							}
						},
						"description": "Safe API error"
					}
				},
				"security": [
					{
						"OAuth2PasswordBearer": []
					}
				],
				"summary": "List Videos By Channel",
				"tags": ["Videos"]
			}
		},
...
	}
}